requirements.txt
imgurpython==1.1.7
Pillow==3.4.2
-e git+https://github.com/JKolios/prawbot.git@master#egg=prawbot
